+++ src/__init__.py
@@ -2,5 +2,5 @@
 metadata for the package
 """
 
-__version__ = "1.10.2"
+__version__ = "1.11.0"
 __author__ = "Edmundo Garcia: ingega@gmail.com"